Add in a view for the cards controller and use jbuilder to render it.
