+++ Gemfile
@@ -9,6 +9,8 @@ gem 'bcrypt'
 
 gem 'sqlite3'
 
+gem 'jbuilder'
+
 group :test do
   gem 'rspec-rails'
   gem "codeclimate-test-reporter", group: :test, require: nil
@@ -24,18 +26,3 @@ group :development, :test do
   gem 'spring'
   gem 'faker'
 end
-
-# To use ActiveModel has_secure_password
-# gem 'bcrypt', '~> 3.1.7'
-
-# To use Jbuilder templates for JSON
-# gem 'jbuilder'
-
-# Use unicorn as the app server
-# gem 'unicorn'
-
-# Deploy with Capistrano
-# gem 'capistrano', :group => :development
-
-# To use debugger
-# gem 'ruby-debug19', :require => 'ruby-debug'
